+++ views/menu/view.php
@@ -33,6 +33,7 @@ $this->params['breadcrumbs'][] = $this->title;
             'menuParent.name:text:Parent',
             'name',
             'route',
+            'order',
         ],
     ]) ?>